requirements.txt
Django==6.0.3
djangorestframework==3.17.0
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
requests==2.31.0
PyJWT==2.12.1
python-dotenv==1.2.2
scikit-fuzzy==0.5.0
numpy==2.4.3
scipy==1.17.1
reportlab==4.4.10
Pillow==10.2.0
gunicorn==25.1.0
whitenoise==6.12.0
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
postgrest==2.28.3
realtime==2.28.3
storage3==2.28.3
httpx==0.28.1
pydantic==2.12.5
pydantic_core==2.41.5
anyio==4.12.1
asgiref==3.11.1
sqlparse==0.5.5
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
websockets==15.0.1
yarl==1.23.0
multidict==6.7.1
propcache==0.4.1
charset-normalizer==3.4.6
annotated-types==0.7.0
typing_extensions==4.15.0
deprecation==2.1.0
packaging==26.0
StrEnum==0.4.15
cachetools==6.2.6
fsspec==2026.2.0
mmh3==5.2.1
pyroaring==1.0.4
strictyaml==1.7.3
tenacity==9.1.4
zstandard==0.25.0
pyiceberg==0.11.1
